moved sort column
